Add World

+++ LogicProject/Potion.cs
@@ -9,11 +9,11 @@ namespace LogicProject
     public class Potion : Item
     {
         public float HealHP { get; set; }
-        public int ReduceMana { get; set; }
+        public float ReduceMana { get; set; }
 
         public Potion() { }
 
-        public Potion(int id, string name, float heal, int reduce) : base(id, name)
+        public Potion(int id, string name, float heal, float reduce) : base(id, name)
         {
             HealHP = heal; ReduceMana = reduce;
         }
